userinfo添加请求modele

+++ iOSStar/Model/ShareDataModel/UserInfoModel.swift
@@ -10,7 +10,7 @@ import UIKit
 class UserInfoModel: NSObject {
     dynamic var nick_name : String = " "
     dynamic var market_cap : Int = 0
-    dynamic var is_setpwd : Int64 = 0
+    dynamic var is_setpwd : Int = 0
     dynamic var balance : Double = 0
     dynamic var total_amt  :Double = 0
     dynamic var head_url : String = " "
